REFACTOR: 1. Switch de responsabilidades de controllers por services y viceversa
