+++ backend/controllers/Controller.ts
@@ -2,13 +2,13 @@ import {Response} from "express";
 import {InvalidArgumentsError} from "../errors/InvalidArgumentsError";
 import {PersistanceError} from "../errors/PersistanceError";
 
-export abstract class Service {
+export abstract class Controller {
     okResponse(res: Response, message: string): void {
-        res.status(200).json({ message });
+        res.status(200).json({ message }).send();
     }
 
     createdResponse(res: Response, message: string): void {
-        res.status(201).json({message});
+        res.status(201).json({message}).send();
     }
 
     noContentResponse(res: Response): void {
@@ -16,19 +16,19 @@ export abstract class Service {
     }
 
     badRequestResponse(res: Response, message: string): void {
-        res.status(400).json({ message });
+        res.status(400).json({ message }).send();
     }
 
     unauthorizedResponse(res: Response, message: string): void {
-        res.status(401).json({ message });
+        res.status(401).json({ message }).send();
     }
 
     notFoundResponse(res: Response, message: string): void {
-        res.status(404).json({ message });
+        res.status(404).json({ message }).send();
     }
 
     internalServerErrorResponse(res: Response, message: string): void {
-        res.status(500).json({ message });
+        res.status(500).json({ message }).send();
     }
 
     handleError(error: Error, res: Response): void {
